LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.strokeRect.transform.worker-expected.txt


PASS fillRect is affected by transforms 
